+++ libharp/harp-utils.c
@@ -188,7 +188,7 @@ int harp_path_find_file(const char *searchpath, const char *filename, char **loc
     char *path_component;
     char *filepath = NULL;
     int filepath_length = 0;
-    int filename_length = strlen(filename);
+    int filename_length = (int)strlen(filename);
 
     if (searchpath == NULL || searchpath[0] == '\0')
     {
@@ -221,7 +221,7 @@ int harp_path_find_file(const char *searchpath, const char *filename, char **loc
             p++;
         }
 
-        path_component_length = strlen(path_component);
+        path_component_length = (int)strlen(path_component);
         if (filepath_length < path_component_length + filename_length + 1)
         {
             char *new_filepath;
@@ -273,8 +273,8 @@ int harp_path_from_path(const char *initialpath, int is_filepath, const char *ap
     int initialpath_length;
     int appendpath_length;
 
-    initialpath_length = strlen(initialpath);
-    appendpath_length = (appendpath == NULL ? 0 : strlen(appendpath));
+    initialpath_length = (int)strlen(initialpath);
+    appendpath_length = (appendpath == NULL ? 0 : (int)strlen(appendpath));
 
     if (is_filepath && initialpath_length > 0)
     {
@@ -354,7 +354,7 @@ int harp_path_for_program(const char *argv0, char **location)
     {
         /* use PATH */
 #ifdef WIN32
-        int argv0_length = strlen(argv0);
+        int argv0_length = (int)strlen(argv0);
 
         if (argv0_length <= 4 || strcmp(&argv0[argv0_length - 4], ".exe") != 0)
         {
